requirements.txt
streamlit
pandas
python-dotenv
google-generativeai
groq
PyPDF2
python-docx
reportlab
matplotlib